components/featurestore/requirements.txt
google-cloud-aiplatform==1.10.0
google-cloud-bigquery==2.32.0
kfp==1.8.11